+++ pdf/test/ttf_test.dart
@@ -20,6 +20,43 @@ import 'dart:typed_data';
 import 'package:pdf/pdf.dart';
 import 'package:test/test.dart';
 
+void printText(PdfGraphics canvas, String text, PdfFont font, double top) {
+  text = text + font.fontName;
+  const double fontSize = 20.0;
+  final PdfFontMetrics metrics = font.stringMetrics(text) * fontSize;
+
+  const double deb = 5.0;
+
+  const double x = 50.0;
+  final double y = canvas.page.pageFormat.height - top;
+
+  canvas
+    ..drawRect(x + metrics.left, y + metrics.top, metrics.width, metrics.height)
+    ..setColor(const PdfColor(0.9, 0.9, 0.9))
+    ..fillPath()
+    ..drawLine(x + metrics.left - deb, y, x + metrics.right + deb, y)
+    ..setColor(PdfColor.blue)
+    ..strokePath()
+    ..drawLine(x + metrics.left - deb, y + metrics.ascent,
+        x + metrics.right + deb, y + metrics.ascent)
+    ..setColor(PdfColor.green)
+    ..strokePath()
+    ..drawLine(x + metrics.left - deb, y + metrics.descent,
+        x + metrics.right + deb, y + metrics.descent)
+    ..setColor(PdfColor.purple)
+    ..strokePath()
+    ..setColor(const PdfColor(0.3, 0.3, 0.3))
+    ..drawString(font, fontSize, text, x, y);
+}
+
+void printTextTtf(PdfGraphics canvas, String text, File ttfFont, double top) {
+  final Uint8List fontData = ttfFont.readAsBytesSync();
+  final PdfTtfFont font =
+      PdfTtfFont(canvas.page.pdfDocument, fontData.buffer.asByteData());
+
+  printText(canvas, text, font, top);
+}
+
 void main() {
   test('Pdf', () {
     final PdfDocument pdf = PdfDocument();
@@ -27,27 +64,11 @@ void main() {
         PdfPage(pdf, pageFormat: const PdfPageFormat(500.0, 300.0));
 
     final PdfGraphics g = page.getGraphics();
-    final Uint8List ttfData = File('open-sans.ttf').readAsBytesSync();
-    final PdfTtfFont ttf = PdfTtfFont(pdf, ttfData.buffer.asByteData());
-    const String s = 'Hello World!';
-    PdfRect r = ttf.stringBounds(s);
-    const double FS = 20.0;
-    const PdfColor pdfColor = PdfColor(0.0, 1.0, 1.0);
-    g.setColor(pdfColor);
-    g.drawRect(50.0 + r.x * FS, 30.0 + r.y * FS, r.width * FS, r.height * FS);
-    g.fillPath();
-    g.setColor(const PdfColor(0.3, 0.3, 0.3));
-    g.drawString(ttf, FS, s, 50.0, 30.0);
+    int top = 0;
+    const String s = 'Hello ';
 
-    final Uint8List robotoData = File('roboto.ttf').readAsBytesSync();
-    final PdfTtfFont roboto = PdfTtfFont(pdf, robotoData.buffer.asByteData());
-
-    r = roboto.stringBounds(s);
-    g.setColor(const PdfColor(0.0, 1.0, 1.0));
-    g.drawRect(50.0 + r.x * FS, 130.0 + r.y * FS, r.width * FS, r.height * FS);
-    g.fillPath();
-    g.setColor(const PdfColor(0.3, 0.3, 0.3));
-    g.drawString(roboto, FS, s, 50.0, 130.0);
+    printTextTtf(g, s, File('open-sans.ttf'), 30.0 + 30.0 * top++);
+    printTextTtf(g, s, File('roboto.ttf'), 30.0 + 30.0 * top++);
 
     final File file = File('ttf.pdf');
     file.writeAsBytesSync(pdf.save());